requirements.txt
pyaudio
google-cloud-speech
google-api-core
grpcio
streamlit

pyaudiowpatch